add cast for some picky systems (reported by KH, not seen on any of mine)
git-svn-id: https://svn.r-project.org/R/trunk@42733 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/nmath/rmultinom.c
@@ -76,7 +76,7 @@ void rmultinom(int n, double* prob, int K, int* rN)
     }
     if(fabs(p_tot - 1.) > 1e-7)
 	MATHLIB_ERROR(_("rbinom: probability sum should be 1, but is %g"), 
-		      p_tot);
+		      (double) p_tot);
     if (n == 0) return;
     if (K == 1 && p_tot == 0.) return;/* trivial border case: do as rbinom */
